backend/requirements-lock.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   httpx
    #   openai
boto3==1.37.22
    # via -r requirements.txt
botocore==1.37.38
    # via
    #   boto3
    #   s3transfer
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.6
    # via requests
cryptography==46.0.5
    # via google-auth
distro==1.9.0
    # via openai
google-auth==2.49.1
    # via google-genai
google-genai==1.10.0
    # via -r requirements.txt
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   google-genai
    #   openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.13.0
    # via openai
jmespath==1.1.0
    # via
    #   boto3
    #   botocore
openai==1.72.0
    # via -r requirements.txt
pillow==11.1.0
    # via -r requirements.txt
pyasn1==0.6.2
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   google-genai
    #   openai
pydantic-core==2.41.5
    # via pydantic
python-dateutil==2.9.0.post0
    # via botocore
requests==2.32.5
    # via
    #   -r requirements.txt
    #   google-genai
s3transfer==0.11.5
    # via boto3
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via openai
tqdm==4.67.3
    # via openai
typing-extensions==4.15.0
    # via
    #   anyio
    #   google-genai
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via
    #   botocore
    #   requests
websockets==15.0.1
    # via google-genai
